+++ Frontend Vehicle/src/App.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Home from './pages/Home/Home';
 import Registration from './pages/Registration/Registration';
-import RegistrationVehicle from './pages/RegistrationVehicle/RegistrationVehicle';
 import Login from './pages/Login/Login';
 import QRpage from './pages/QRpage/QRpage';
 import Navbar from './components/Navbar';
@@ -14,7 +13,6 @@ function App() {
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/register" element={<Registration />} />
-        <Route path="/registration-vehicle" element={<RegistrationVehicle />} />
         <Route path="/login" element={<Login />} />
         <Route path='/qr' element={<QRpage />} />
         <Route path="*" element={<h2>Page Not Found</h2>} />
