+++ Frontend Vehicle/src/pages/Registration/Registration.css
@@ -1,167 +1,199 @@
-.form-section {
-    width: 100%;
-    padding: 25px;
-    font-family: Arial, sans-serif;
-    text-align: center;
-    border: 1px solid #eee; /* Lighter border */
-    border-radius: 8px;
-    background-color: white;
-    margin-bottom: 25px;
-    margin-top: 0; /* Remove top margin from here as it's handled by registration-form */
-    height: auto; /* Allow height to adjust based on content */
-  }
-  
-  .form-section h1 {
-    font-size: 28px;
-    font-weight: bold;
-    margin: 10px 0 20px;
-    padding: 10px;
-    color: #333;
-    width: 100%;
-    text-align: center;
-  }
-  
-  .form-section h2 {
-    text-align: left;
-    margin: 20px 0 15px;
-    color: #444;
-  }
-  
-  .form-section p {
-    margin-bottom: 20px;
-    text-align: center;
-  }
-  
-  .form-section a {
-    color: #007bff;
-    text-decoration: underline;
-    cursor: pointer;
-  }
-  
-  .registration-container {
-    overflow: visible;
-  }
-  
-  form {
-    max-width: 800px;
-    margin: 80px auto 40px; /* Increased top margin from 40px to 80px */
-    padding: 30px;
-    background-color: white; /* Changed from #f9f9f9 to white for better contrast */
-    border-radius: 10px;
-    box-shadow: 0 0 15px rgba(0, 0, 0, 0.1);
-    display: flex;
-    flex-direction: column;
-    align-items: center; /* Center align the form contents */
-  }
-  
-  .form-group {
-    margin-bottom: 20px; /* Increased spacing */
-    text-align: left;
-    display: flex;
-    flex-direction: column;
-    width: 100%; /* Ensure full width */
-    align-items: flex-start;
-  }
-  
-  .form-group label {
-    display: block;
-    font-weight: bold;
-    margin-bottom: 5px;
-  }
-  
-  .form-group .input-with-button {
-    display: flex;
-    align-items: center;
-    width: 100%;
-    gap: 10px;
-  }
-  
-  .form-group .input-with-button input {
-    flex: 1;
-    padding: 8px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    margin-right: 10px;
-    flex-grow: 1;
-  }
-  
-  .form-group .input-with-button button {
-    padding: 8px 12px;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    min-width: 100px;
-    height: 40px;
-  }
-  
-  .form-group input {
-    width: 100%;
-    padding: 8px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-  }
-  
-  .form-group button {
-    padding: 8px 12px;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-  }
-  
-  .form-group button:hover {
-    background-color: #0056b3;
-  }
-  
-  .next-button {
-    width: 40%;
-    padding: 12px; /* Slightly larger padding */
-    background-color: #ffa500;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    font-size: 16px;
-    cursor: pointer;
-    margin: 20px 0 10px auto; /* Align to the right */
-    display: block;
-  }
-  
-  .next-button:hover {
-    background-color: #cc8400;
-  }
+/* Merged styles from Registration.css and RegistrationVehicle.css */
+.form-section, .registration-vehicle-container {
+  width: 100%;
+  padding: 25px;
+  font-family: Arial, sans-serif;
+  text-align: center;
+  border: 1px solid #eee; /* Lighter border */
+  border-radius: 8px;
+  background-color: #f9f9f9;
+  margin-bottom: 25px;
+  margin-top: 120px; /* Adjusted top margin to account for navbar height */
+  height: auto; /* Allow height to adjust based on content */
+}
 
-  .registration-form {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    max-width: 800px;
-    margin: 100px auto 40px; /* Increased top margin to account for navbar */
-    padding-top: 30px; /* Add padding to top */
-  }
-  
-  .page-title {
-    font-size: 28px;
-    font-weight: bold;
-    text-align: center;
-    width: 100%;
-    margin: 10px 0 20px;
-    padding: 10px;
-    color: #333;
-  }
-  
-  .login-link {
-    text-align: center;
-    width: 100%;
-    margin-bottom: 20px;
-  }
-  
-  .section-title {
-    text-align: left;
-    width: 100%;
-    margin: 20px 0 15px;
-    color: #444;
-  }
+h1 {
+  text-align: center;
+  margin-bottom: 25px;
+  color: #333;
+  font-size: 28px;
+  font-weight: bold;
+}
+
+h2 {
+  text-align: left;
+  margin: 20px 0 15px;
+  color: #444;
+}
+
+p {
+  text-align: center;
+  margin-bottom: 20px;
+  color: #555;
+}
+
+a {
+  color: #007bff;
+  text-decoration: underline;
+  cursor: pointer;
+}
+
+.registration-container {
+  overflow: visible;
+}
+
+form {
+  max-width: 800px;
+  margin: 80px auto 40px; /* Increased top margin from 40px to 80px */
+  padding: 30px;
+  background-color: white; /* Changed from #f9f9f9 to white for better contrast */
+  border-radius: 10px;
+  box-shadow: 0 0 15px rgba(0, 0, 0, 0.1);
+  display: flex;
+  flex-direction: column;
+  align-items: center; /* Center align the form contents */
+}
+
+.form-group {
+  margin-bottom: 20px; /* Increased spacing */
+  text-align: left;
+  display: flex;
+  flex-direction: column;
+  width: 100%; /* Ensure full width */
+}
+
+.form-group label {
+  display: block;
+  font-weight: bold;
+  margin-bottom: 5px;
+}
+
+.form-group .input-with-button {
+  display: flex;
+  align-items: center;
+}
+
+.form-group .input-with-button input {
+  flex: 1;
+  padding: 8px;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  margin-right: 10px;
+}
+
+input,
+select {
+  width: 100%;
+  padding: 10px 12px;
+  margin-bottom: 8px;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  font-size: 16px;
+  transition: border-color 0.3s;
+}
+
+input:focus,
+select:focus {
+  outline: none;
+  border-color: #4caf50;
+  box-shadow: 0 0 5px rgba(76, 175, 80, 0.2);
+}
+
+/* Vehicle number styles */
+.vehicle-number-inputs {
+  display: flex;
+  gap: 15px;
+  align-items: center;
+}
+
+/* Fuel type options */
+.fuel-type-options {
+  display: flex;
+  justify-content: space-around;
+  padding: 10px 0;
+}
+
+.fuel-type-options label {
+  display: inline-flex;
+  align-items: center;
+  margin-right: 20px;
+}
+
+.fuel-type-options input[type="radio"] {
+  width: auto;
+  margin-right: 5px;
+}
+
+.form-group button {
+  padding: 8px 12px;
+  background-color: #007bff;
+  color: white;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+}
+
+.form-group button:hover {
+  background-color: #0056b3;
+}
+
+/* Button styles */
+.next-button {
+  width: 40%;
+  padding: 12px; /* Slightly larger padding */
+  background-color: #ffa500;
+  color: white;
+  border: none;
+  border-radius: 4px;
+  font-size: 16px;
+  cursor: pointer;
+  margin: 20px 0 10px auto; /* Align to the right */
+  display: block;
+}
+
+.next-button:hover {
+  background-color: #cc8400;
+}
+
+.form-footer {
+  text-align: left;
+  margin-top: 25px;
+}
+
+.form-buttons {
+  display: flex;
+  justify-content: space-between;
+  margin-top: 15px;
+  gap: 20px;
+}
+
+.back-button,
+.register-button {
+  padding: 12px 25px;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  font-size: 16px;
+  font-weight: 600;
+  transition: background-color 0.3s, transform 0.1s;
+  flex: 1;
+}
+
+.back-button {
+  background-color: #f0f0f0;
+  color: #555;
+}
+
+.back-button:hover {
+  background-color: #e0e0e0;
+}
+
+.register-button {
+  background-color: #4caf50;
+  color: white;
+}
+
+.register-button:hover {
+  background-color: #45a049;
+  transform: translateY(-1px);
+}
